+++ graphene_django_crud/__init__.py
@@ -2,7 +2,7 @@
 from graphene.pyutils.version import get_version
 
 from .types import DjangoGrapheneCRUD
-
+from .types import apply_where as where_input_to_queryset_filter_args
 
 
 __version__ = "1.0.3"
@@ -10,4 +10,6 @@ __version__ = "1.0.3"
 __all__ = (
     "__version__",
     "DjangoGrapheneCRUD",
+    "where_input_to_queryset_filter_args"
+
 )