+++ graphene_django_crud/types.py
@@ -213,6 +213,17 @@ class DjangoGrapheneCRUD(graphene.ObjectType):
     def Type(cls):
         return cls
 
+    @classmethod
+    def WhereInputType(cls, *args, **kwargs):
+        return convert_model_to_input_type(cls._meta.model, input_flag="where", registry=cls._meta.registry)
+
+    @classmethod
+    def CreateInputType(cls, *args, **kwargs):
+        return convert_model_to_input_type(cls._meta.model, input_flag="create", registry=cls._meta.registry)
+
+    @classmethod
+    def UpdateInputType(cls, *args, **kwargs):
+        return convert_model_to_input_type(cls._meta.model, input_flag="update", registry=cls._meta.registry)
 
     @classmethod
     def ReadField(cls, *args, **kwargs):
@@ -220,7 +231,7 @@ class DjangoGrapheneCRUD(graphene.ObjectType):
         arguments = OrderedDict()
         arguments.update({
             "where": graphene.Argument(
-                convert_model_to_input_type(cls._meta.model, input_flag="where_with_operator", registry=cls._meta.registry),
+                convert_model_to_input_type(cls._meta.model, input_flag="where", registry=cls._meta.registry),
                 required=True
             ),
         })
@@ -243,7 +254,7 @@ class DjangoGrapheneCRUD(graphene.ObjectType):
 
         arguments = OrderedDict()
         arguments.update({
-            "where": graphene.Argument(convert_model_to_input_type(cls._meta.model, input_flag="where_with_operator", registry=cls._meta.registry)),
+            "where": graphene.Argument(convert_model_to_input_type(cls._meta.model, input_flag="where", registry=cls._meta.registry)),
             "limit": graphene.Int(),
             "offset" : graphene.Int(),
             "orderBy" : graphene.List(graphene.String)